+++ components/cart/button/CartSelectButton.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import Button from '../Button';
+import Button from '@/components/Button';
 import { FaCheck } from 'react-icons/fa';
 
 export default function CartSelectButton() {
