+++ components/cart/modal/CartModal.module.css
@@ -44,3 +44,16 @@
   font-size: 24px;
   font-weight: 700;
 }
+
+.bar {
+  width: 30px;
+  height: 5px;
+  border-radius: 10px;
+  background-color: var(--color-lightgrey);
+  position: absolute;
+  top: 10px;
+  left: 50%;
+  transform: translateX(-50%);
+  cursor: pointer;
+  -webkit-user-drag: none;
+}
