+++ pages/cart.tsx
@@ -1,5 +1,5 @@
-import CartList from '@/components/cart/CartList';
-import CartResult from '@/components/cart/CartResult';
+import CartList from '@/components/cart/list/CartList';
+import CartResult from '@/components/cart/result/CartResult';
 import styles from '@/styles/Cart.module.css';
 export default function Cart() {
   return (
